move server to its own package
